Exports, Uploads, Caches implemented

+++ src/services/postgres/PlaylistsService.js
@@ -5,9 +5,10 @@ const InvariantError = require('../../exceptions/InvariantError')
 const NotFoundError = require('../../exceptions/NotFoundError')
 
 class PlaylistsService {
-    constructor (collaborationsService) {
+    constructor (collaborationsService, cacheService) {
         this._pool = new Pool()
         this._collaborationsService = collaborationsService
+        this._cacheService = cacheService
     }
 
     async addPlaylist ({ name, owner }) {
@@ -24,25 +25,33 @@ class PlaylistsService {
             throw new InvariantError('Gagal menambahkan playlist!')
         }
 
+        await this._cacheService.delete(`playlists:${owner}`)
         return result.rows[0].id
     }
 
     async getPlaylists (userId) {
-        const query = {
-            text: `SELECT playlists.id, playlists.name, users.username FROM playlists
-            INNER JOIN users ON playlists.owner = users.id
-            LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-            WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
-            values: [userId]
-        }
+        try {
+            const result = await this._cacheService.get(`playlists:${userId}`)
+            return JSON.parse(result)
+        } catch (err) {
+            const query = {
+                text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+                INNER JOIN users ON playlists.owner = users.id
+                LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+                WHERE playlists.owner = $1 OR collaborations.user_id = $1`,
+                values: [userId]
+            }
 
-        const result = await this._pool.query(query)
-        return result.rows
+            const result = await this._pool.query(query)
+
+            await this._cacheService.set(`playlists:${userId}`, JSON.stringify(result.rows))
+            return result.rows
+        }
     }
 
     async delPlaylistById (playlistId) {
         const query = {
-            text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+            text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
             values: [playlistId]
         }
 
@@ -51,6 +60,9 @@ class PlaylistsService {
         if (!result.rowCount) {
             throw new NotFoundError('Playlist gagal dihapus. Id tidak ditemukan')
         }
+
+        const { owner } = result.rows[0]
+        await this._cacheService.delete(`playlists:${owner}`)
     }
 
     async addSongtoPlaylist (playlistId, songId) {
@@ -67,25 +79,29 @@ class PlaylistsService {
             throw new InvariantError('Gagal menambahkan lagu ke playlist.')
         }
 
+        await this._cacheService.delete(`songs:${playlistId}`)
         return result.rows[0].id
     }
 
     async getSongsFromPlaylist (playlistId) {
-        const query = {
-            text: `SELECT songs.id, songs.title, songs.performer FROM playlists
-            INNER JOIN playlistsongs ON playlistsongs.playlist_id = playlists.id
-            INNER JOIN songs ON songs.id = playlistsongs.song_id
-            WHERE playlists.id = $1`,
-            values: [playlistId]
+        try {
+            const result = await this._cacheService.get(`songs:${playlistId}`)
+            return JSON.parse(result)
+        } catch (err) {
+            const query = {
+                text: `SELECT songs.id, songs.title, songs.performer FROM playlists
+                INNER JOIN playlistsongs ON playlistsongs.playlist_id = playlists.id
+                INNER JOIN songs ON songs.id = playlistsongs.song_id
+                WHERE playlists.id = $1`,
+                values: [playlistId]
+            }
+
+            const result = await this._pool.query(query)
+
+            await this._cacheService.set(`songs:${playlistId}`, JSON.stringify(result.rows))
+
+            return result.rows
         }
-
-        const result = await this._pool.query(query)
-
-        if (!result.rowCount) {
-            throw new InvariantError('Gagal menampilkan lagu pada playlist.')
-        }
-
-        return result.rows
     }
 
     async delSongFromPlaylist (playlistId, songId) {
@@ -99,6 +115,8 @@ class PlaylistsService {
         if (!result.rowCount) {
             throw new InvariantError('Lagu gagal dihapus dari playlist.')
         }
+
+        await this._cacheService.delete(`songs:${playlistId}`)
     }
 
     async verifyPlaylistOwner (playlistId, userId) {
